#18: minor code refactoring
